gameplay: fix deploy times bugs
switch from bonus is only applied if the weapon we are switching
from has fully deployed. however, the calculation for "fully deployed"
was incorrect. it used the default switch time of 0.5, instead of also
taking into account weapon attributes which affect deploy time.

this affected the degreaser most notably, which would not
allow pyros to switch to their flare gun with increased switch speed
as soon as the degreaser deployed (0.33s).
instead, players would have to wait until 0.5s had passed before switching
which is unintuitive

also, with this change, we can easily fix bugs with some weapons
which did not consider the weapon being fully deployed
before applying a buff. some of these may be controversial so
feel free to ignore them, they are just here as examples to demonstrate
the new functionality. note that the winger nerf would be a requested
change by the competitive community.

finally, fix a bug where swapping to another weapon immediately after
would erase the deploy time penalty/bonus. this would penalize
mouse wheel users when using the degreaser for the switch speed bonus,
and it also can be exploited to bypass holster penalties for items
like the fists of steel.

ref: ValveSoftware/Source-1-Games#3488
ref: ValveSoftware/Source-1-Games#2125