gameplay: movement speed modifiers were incorrect
stunned movement was not applying correctly to diagonal movement
because diagonal movement is a combination of two axes
they exceed max speed until clamped by CheckParameters

stunned movement slows were applying before this clamp,
so they did not properly slow the actual speed of the player
which could be much lower than the unclamped speed

in addition, high max speed boosts (like the Baby Face's Blaster)
did not apply unless the input speed was default speed
since stunned movement manipulates input speed directly
instead of changing player velocity, this broke the input speed boost
and thus scout's max speed would be capped higher but not have the
input speed to drive it.

with these changes, a 0.6 slow (as applied by the Natascha at close range),
slows down a full charge BFB Scout from 520 to 208 in all directions (40% of 520),
as intended and proper! And of course, normal Scout speed is slowed from
400 to 160 in all directions (40% of 400).

however, with this fix, the slow amount is drastically more effective
than it was before. so, we apply legacy handling to reduce slow amount
by a percentage value relative to the old effective scaling.

a scaling value was chosen to get a near average between the effective slow
applied with a 100% slow, and the full slow amount applied with the fix

note that there is still technically an issue with strafe tapping
boosting your movement speed, but this is a game movement mechanic which
I don't think is related to stuns, which seem to be intended to apply
to movement controls, not exactly correlating to output velocity.
or else stuns would not be a different movement mechanic vs. percentage
movement speed slows.

ref: ValveSoftware/Source-1-Games#3721